[REFACTOR] Certik Audit Recommendation - Exhibit 4
Description:
The specified lines contain a multiplication of a “uint256” value with the contract member
“_gasPriceScaling” followed by a division with the constant value
“GAS_PRICE_SCALING_SCALE”.

Recommendations:
It is possible that the multiplication will overflow, causing incorrect values to be passed as the
value to charge as a fee for the transaction since the variable “_gasPriceScaling” relies on user
input. We advise the utilization of the already-imported “SafeMath” library to conduct the
multiplication to ensure no overflow occurs however unlikely.