[REFACTOR] Certik Audit Recommendation - Exhibit 5
Description:
The specified lines utilize the “super” specifier to invoke a contract, however the “ERC20Fees”
utilizes multiple inheritance.

Recommendations:
In Solidity you read class-name declarations right-to-left from lowest-to-highest in contrast to
other programming languages such as Python where the opposite holds true. Thus, the “super”
keyword may be ambiguous / confusing to a reader of the codebase. Although the code
compiles correctly, the “super” keyword should be replaced with the name of the contract the
function belongs to to aid in the legibility of the codebase.