[REFACTOR] Certik Audit Recommendation - Exhibit 6
Description:
Line 8 of ERC20Full.sol contains the expression “2**256 - 1” which is equivalent to the
maximum value a “uint256” type can hold.

Recommendations:
This value would otherwise overflow as “2**256” cannot be represented in the “uint256” limit.
The compiler simply swaps “2**256 - 1” with the evaluated result, however this may lead readers
of the codebase to assume it is possible to conduct such operations. We advise the utilization
of a proper Solidity expression such as “~uint256(0)”, which inverts all the bits of the number
zero represented in the form of a “uint256”.