[REFACTOR] Certik Audit Recommendation - Exhibit 7
Description:
The “transferFrom” function contains statements that “bypass the internal allowance
manipulation and checks for the whitelisted operator (i.e. _msgSender())” as stated in the
comments.

Recommendations:
The “if” clause at L44 checks whether the caller of the “transferFrom” function is the “Operator”.
If it is, the function proceeds by transferring the funds between the two accounts provided as
input to the function regardless of whether the accounts have approved the “Operator”.
Although this is a side-effect of what the contract is meant to achieve, utmost care should be
taken when handling the private keys of the “Operator” as it allows one who holds the private
key of an “Operator” to drain any account of token funds at will.